requirements.txt
absl-py==2.0.0
ace_tools==0.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.3
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.9.0
branca==0.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.9.0
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.115.12
fastjsonschema==2.19.0
filelock==3.18.0
findspark==2.0.1
flatbuffers==23.5.26
folium==0.15.0
fonttools==4.46.0
fsspec==2025.3.2
gast==0.5.4
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.6
imageio==2.37.0
imbalanced-learn==0.12.3
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jiter==0.8.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
Mesa==3.0.3
ml-dtypes==0.2.0
mpmath==1.3.0
narwhals==1.37.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.4.2
nltk==3.8.1
numba==0.61.0
numpy==1.26.2
oauthlib==3.2.2
openai==1.57.4
opencv-python==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
optuna==4.3.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.0.0
plotly==5.18.0
pmdarima==2.0.4
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9
pyarrow==20.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.1
pyspark==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==25.1.1
referencing==0.31.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.2
rsa==4.9
safetensors==0.5.3
scikit-learn==1.3.2
scikit-surprise==1.1.3
scipy==1.11.4
seaborn==0.13.2
sentence-transformers==4.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.1
streamlit==1.45.0
streamlit-autorefresh==1.0.1
sympy==1.13.1
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.9.0
tkhtmlview==0.2.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.0
transformers==4.51.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.12
Werkzeug==3.0.1
wordcloud==1.9.2
wrapt==1.14.1
xgboost==2.1.1
